+++ reports/certificates.php
@@ -18,7 +18,7 @@
  * Certificates report page for block_reports_custom.
  *
  * @package    block_reports_custom
- * @copyright  2024 Your Organization
+ * @copyright  2025 Alonso Arias <soporte@ingeweb.co>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -132,9 +132,11 @@ $PAGE->set_title(get_string('certificates_report', 'block_reports_custom'));
 $PAGE->set_heading(get_string('certificates_report', 'block_reports_custom'));
 $PAGE->set_pagelayout('report');
 
-// Include JavaScript.
-$PAGE->requires->jquery();
-$PAGE->requires->js(new moodle_url('/blocks/reports_custom/reports/certificates.js'));
+// Include AMD JavaScript module.
+$PAGE->requires->js_call_amd('block_reports_custom/certificates', 'init', [[
+    'formSelector' => '#filtersForm',
+    'reportContainerSelector' => '#reportData',
+]]);
 
 // Get total count and paginated records.
 $totalcount = block_reports_custom_count_certificates_records($filters);